Add legacy between the different king of window
